+++ HTTPShortcuts/app/src/main/kotlin/ch/rmy/android/http_shortcuts/activities/categories/editor/models/CategoryBackground.kt
@@ -1,0 +1,7 @@
+package ch.rmy.android.http_shortcuts.activities.categories.editor.models
+
+enum class CategoryBackground {
+    DEFAULT,
+    COLOR,
+    WALLPAPER,
+}