+++ HTTPShortcuts/app/src/main/kotlin/ch/rmy/android/http_shortcuts/data/migration/DatabaseMigration.kt
@@ -1,6 +1,7 @@
 package ch.rmy.android.http_shortcuts.data.migration
 
 import ch.rmy.android.framework.utils.UUIDUtils.newUUID
+import ch.rmy.android.http_shortcuts.data.migration.migrations.CategoryBackgroundMigration
 import ch.rmy.android.http_shortcuts.data.migration.migrations.ParameterTypeMigration
 import ch.rmy.android.http_shortcuts.data.migration.migrations.RemoveLegacyActionsMigration
 import ch.rmy.android.http_shortcuts.data.migration.migrations.ReplaceActionsWithScriptsMigration
@@ -385,6 +386,7 @@ class DatabaseMigration : RealmMigration {
                 schema.get("Category")!!
                     .addField("shortcutClickBehavior", String::class.java)
             }
+            50L -> CategoryBackgroundMigration.migrateRealm(realm)
             else -> throw IllegalArgumentException("Missing migration for version $newVersion")
         }
         updateVersionNumber(realm, newVersion)
@@ -410,6 +412,6 @@ class DatabaseMigration : RealmMigration {
 
     companion object {
 
-        const val VERSION = 49L
+        const val VERSION = 50L
     }
 }