+++ HTTPShortcuts/app/src/main/kotlin/ch/rmy/android/http_shortcuts/data/migration/ImportMigrator.kt
@@ -1,6 +1,7 @@
 package ch.rmy.android.http_shortcuts.data.migration
 
 import ch.rmy.android.framework.utils.UUIDUtils.newUUID
+import ch.rmy.android.http_shortcuts.data.migration.migrations.CategoryBackgroundMigration
 import ch.rmy.android.http_shortcuts.data.migration.migrations.ParameterTypeMigration
 import ch.rmy.android.http_shortcuts.data.migration.migrations.RemoveLegacyActionsMigration
 import ch.rmy.android.http_shortcuts.data.migration.migrations.ReplaceActionsWithScriptsMigration
@@ -142,6 +143,9 @@ internal object ImportMigrator {
                         }
                     }
                 }
+            }
+            50L -> { // 2.15.0
+                CategoryBackgroundMigration.migrateImport(base)
             }
         }
     }