+++ HTTPShortcuts/app/src/main/kotlin/ch/rmy/android/http_shortcuts/data/migration/migrations/CategoryBackgroundMigration.kt
@@ -1,0 +1,32 @@
+package ch.rmy.android.http_shortcuts.data.migration.migrations
+
+import com.google.gson.JsonObject
+import io.realm.DynamicRealm
+
+object CategoryBackgroundMigration : BaseMigration {
+
+    override fun migrateImport(base: JsonObject) {
+        base.getAsJsonArray("categories")
+            ?.map { it.asJsonObject }
+            ?.forEach { category ->
+                when (category.get("background")?.takeUnless { it.isJsonNull }?.asString) {
+                    "white" -> category.addProperty("background", "default")
+                    "black" -> category.addProperty("background", "color=#000000")
+                }
+            }
+    }
+
+    override fun migrateRealm(realm: DynamicRealm) {
+        realm.where("Category")
+            .findAll()
+            .forEach { category ->
+                when (category.getString("background")) {
+                    "white" -> category.setString("background", "default")
+                    "black" -> category.setString("background", "color=#000000")
+                }
+            }
+    }
+
+    override val version: Int
+        get() = 50
+}